Finishing the rebase and fixing some conflicts
