+++ src/components/section-header/index.tsx
@@ -1,10 +1,14 @@
 import React from "react";
-import "./style.css";
+import { LogoutButton } from "../logout-button";
+import * as S from "./style";
 
 export const SectionHeader: React.FC = () => {
+  const loggedIn = window.localStorage.getItem("auth-token");
+
   return (
-    <div className="title">
-      <h1>Bem-vindo(a) à Taqtile!</h1>
-    </div>
+    <S.HeaderWrapper>
+      <S.HeaderTitle>Bem-vindo(a) à Taqtile!</S.HeaderTitle>
+      {loggedIn && <LogoutButton />}
+    </S.HeaderWrapper>
   );
 };
