--- src/components/section-header/style.css
@@ -1,23 +1,0 @@
-@import "../../global.css";
-
-.title {
-  z-index: 100;
-  width: 100vw;
-  height: 20vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  position: fixed;
-  top: 0;
-  font-size: 2vh;
-  color: var(--light-text-color);
-  background-color: var(--bg-dark-color);
-  border-radius: 0 0 6vh 6vh;
-}
-
-@media (max-width: 768px) {
-  .title {
-    border-radius: 0 0 4vh 4vh;
-  }
-}
